Update spinchain.py
modified to require min frames seen before allowing data to be accepted to prevent mislabel matches, also changed from id to name in case of object detection loss creating a new id forcing the match to end due to 3 second of lost id and an update that allows the match to end if the other bey is still moving if the first stops.  also telemetry, and merkle tree, and handshake added for trustlessness.